Fix failure when running some Alembic's commands
For instance, `history` command failed with:

```
Can't invoke function 'get_bind', as the proxy object has not yet been
established for the Alembic 'Operations' class. Try placing this code
inside a callable.
```

It was due to use of `op` outside migration functions.

For reference, this PR solved the same kind of problem:
https://github.com/apache/incubator-superset/pull/3948/files